EP-3083: Tehty navigationBuilderPublicLinkit builderista uusi default builder
