+++ eperusteet/eperusteet-service/src/main/java/fi/vm/sade/eperusteet/service/impl/navigationpublic/NavigationBuilderPublicDefault.java
@@ -9,44 +9,40 @@ import fi.vm.sade.eperusteet.dto.peruste.PerusteenOsaDto;
 import fi.vm.sade.eperusteet.dto.peruste.PerusteenOsaViiteDto;
 import fi.vm.sade.eperusteet.dto.peruste.PerusteenSisaltoDto;
 import fi.vm.sade.eperusteet.dto.peruste.TekstiKappaleDto;
+import fi.vm.sade.eperusteet.dto.tuva.KoulutuksenOsaDto;
+import fi.vm.sade.eperusteet.dto.vst.KotoKielitaitotasoDto;
+import fi.vm.sade.eperusteet.dto.vst.KotoLaajaAlainenOsaaminenDto;
+import fi.vm.sade.eperusteet.dto.vst.KotoOpintoDto;
+import fi.vm.sade.eperusteet.dto.vst.OpintokokonaisuusDto;
 import fi.vm.sade.eperusteet.service.NavigationBuilderPublic;
 import fi.vm.sade.eperusteet.service.PerusteService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
 
 import java.util.Collections;
+import java.util.List;
 import java.util.Objects;
 import java.util.Set;
 import java.util.stream.Collectors;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Component;
-import org.springframework.transaction.annotation.Transactional;
 
 @Component
-@Transactional
 public class NavigationBuilderPublicDefault implements NavigationBuilderPublic {
 
+    private final PerusteService perusteService;
+
     @Autowired
-    private PerusteService perusteService;
+    public NavigationBuilderPublicDefault(PerusteService perusteService) {
+        this.perusteService = perusteService;
+    }
 
     @Override
     public Set<KoulutustyyppiToteutus> getTyypit() {
         return Collections.emptySet();
     }
 
-    private NavigationNodeDto constructNavigation(PerusteenOsaViiteDto.Laaja sisalto) {
-        NavigationType type = NavigationType.viite;
+    public NavigationNodeDto constructNavigation(PerusteenOsaViiteDto.Laaja sisalto) {
         PerusteenOsaDto.Laaja po = sisalto.getPerusteenOsa();
-        if (po != null) {
-            if (po instanceof TekstiKappaleDto && ((TekstiKappaleDto) po).getLiite() != null && ((TekstiKappaleDto) po).getLiite()) {
-                type = NavigationType.liite;
-            } else if (po instanceof TekstiKappaleDto) {
-                TekstiKappaleDto tk = (TekstiKappaleDto) po;
-                if (PerusteenOsaTunniste.RAKENNE.equals(tk.getTunniste())) {
-                    type = NavigationType.muodostuminen;
-                }
-            } else {
-                type = po.getNavigationType();
-            }
-        }
+        NavigationType type = getNavigationType(po, sisalto.getLapset());
 
         NavigationNodeDto result = NavigationNodeDto
                 .of(type, getPerusteenOsaNimi(sisalto.getPerusteenOsa()), sisalto.getId())
@@ -58,6 +54,52 @@ public class NavigationBuilderPublicDefault implements NavigationBuilderPublic {
         return result;
     }
 
+    private NavigationType getNavigationType(PerusteenOsaDto.Laaja po, List<PerusteenOsaViiteDto.Laaja> lapset) {
+        NavigationType type = NavigationType.viite;
+        if (po == null) {
+            return type;
+        }
+
+        if (isTekstikappaleLiite(po)) {
+            return NavigationType.liite;
+        }
+
+        if (lapset.stream().anyMatch(this::isLinkkisivuType)) {
+            return NavigationType.linkkisivu;
+        }
+
+        if (!(po instanceof TekstiKappaleDto)) {
+            return po.getNavigationType();
+        }
+
+        TekstiKappaleDto tk = (TekstiKappaleDto) po;
+        if (PerusteenOsaTunniste.RAKENNE.equals(tk.getTunniste())) {
+            return NavigationType.muodostuminen;
+        }
+
+        return type;
+    }
+
+    private boolean isTekstikappaleLiite(PerusteenOsaDto.Laaja po) {
+        return po instanceof TekstiKappaleDto && ((TekstiKappaleDto) po).getLiite() != null && ((TekstiKappaleDto) po).getLiite();
+    }
+
+    /**
+     * Jos Navigaationoden yksikin lapsi on tiettyä ennalta määritettyä tyyppiä,
+     * laitetaan parent noden tyypiksi linkkisivu.
+     */
+    private boolean isLinkkisivuType(PerusteenOsaViiteDto.Laaja lapsi) {
+        if (lapsi.getPerusteenOsa() == null) {
+            return false;
+        }
+
+        return lapsi.getPerusteenOsa() instanceof KotoKielitaitotasoDto ||
+               lapsi.getPerusteenOsa() instanceof KotoOpintoDto ||
+               lapsi.getPerusteenOsa() instanceof KotoLaajaAlainenOsaaminenDto ||
+               lapsi.getPerusteenOsa() instanceof OpintokokonaisuusDto ||
+               lapsi.getPerusteenOsa() instanceof KoulutuksenOsaDto;
+    }
+
     @Override
     public NavigationNodeDto buildNavigation(Long perusteId, String kieli) {
         PerusteKaikkiDto peruste = perusteService.getJulkaistuSisalto(perusteId);
@@ -65,15 +107,20 @@ public class NavigationBuilderPublicDefault implements NavigationBuilderPublic {
         Set<PerusteenSisaltoDto> sisallot = peruste.getSisallot();
         NavigationNodeDto result = NavigationNodeDto.of(NavigationType.root);
 
-        if (sisallot.size() > 0) {
-            PerusteenSisaltoDto sisalto = sisallot.iterator().next();
-            if (sisalto != null) {
-                PerusteenOsaViiteDto.Laaja sisaltoViite = sisalto.getSisalto();
-                if (sisaltoViite != null) {
-                    result.addAll(constructNavigation(sisaltoViite));
-                }
-            }
+        if (sisallot.isEmpty()) {
+            return result;
         }
+
+        PerusteenSisaltoDto sisalto = sisallot.iterator().next();
+        if (sisalto == null) {
+            return result;
+        }
+
+        PerusteenOsaViiteDto.Laaja sisaltoViite = sisalto.getSisalto();
+        if (sisaltoViite != null) {
+            result.addAll(constructNavigation(sisaltoViite));
+        }
+
         return result;
     }
 }
